style: 코드 형식 및 네이밍 변경
