+++ iOS/issue-tracker/issue-tracker/Network/DTO/LabelListResponseDTO.swift
@@ -8,4 +8,5 @@ struct LabelListResponseDTO: Decodable {
     func toDomain() -> LabelList {
         return .init(labels: data.map { DetailLabel(id: $0.id, title: $0.title, content: $0.content ?? EmptyCase.content.description, color: $0.color) })
     }
+    
 }
