+++ iOS/issue-tracker/issue-tracker/Network/DTO/RequestBodyDTO.swift
@@ -1,9 +1,11 @@
 import Foundation
 
 struct IssueIDsDTO: Encodable {
+    
     let issueIds: [Int]
     
     enum CodingKeys: String, CodingKey {
         case issueIds = "issue_ids"
     }
+    
 }
