+++ iOS/issue-tracker/issue-tracker/Network/DTO/ResponseBodyDTO.swift
@@ -1,6 +1,8 @@
 import Foundation
 
 struct ResponseBodyDTO: Decodable {
+    
     let data: String?
     let error: String?
+    
 }
