+++ iOS/issue-tracker/issue-tracker/Network/RequestManager.swift
@@ -32,9 +32,11 @@ class RequestManager {
 }
 
 enum HttpMethod: String {
+    
     case get = "GET"
     case post = "POST"
     case delete = "DELETE"
     case put = "PUT"
     case patch = "PATCH"
+    
 }
