+++ iOS/issue-tracker/issue-tracker/UseCase/DefaultIssueListUseCase.swift
@@ -79,4 +79,5 @@ class DefaultIssueListUseCase: IssueListUseCase {
                 }
             }.store(in: &subscriptions)
     }
+    
 }
