+++ iOS/issue-tracker/issue-tracker/UseCase/LoginUseCase.swift
@@ -4,28 +4,28 @@ import Combine
 class LoginUseCase {
     
     private let jwtManager: JWTManageable
-    private let loginManager: LoginHelper
+    private let loginHelper: LoginHelper
     private let networkManager: NetworkManager
     private var subscriptions: Set<AnyCancellable>
     
     init() {
         self.jwtManager = JWTManager()
-        self.loginManager = LoginHelper()
+        self.loginHelper = LoginHelper()
         self.networkManager = NetworkManager(requestManager: RequestManager(jwtManager: jwtManager), session: URLSession.shared)
         self.subscriptions = Set<AnyCancellable>()
     }
 
     func gitHubLoginURL() -> URL? {
-        return loginManager.getGitHubLoginURL()
+        return loginHelper.getGitHubLoginURL()
     }
     
     func callbackURLscheme() -> String {
-        return loginManager.getCallbackURLscheme()
+        return loginHelper.getCallbackURLscheme()
     }
     
     func convertToGitHubLogInURL(from callbackURL: URL) -> URL? {
-        let code = self.loginManager.extractedAuthorizationCode(from: callbackURL)
-        return self.loginManager.convertedToURL(with: code)
+        let code = self.loginHelper.extractedAuthorizationCode(from: callbackURL)
+        return self.loginHelper.convertedToURL(with: code)
     }
     
     func executeGitHubLogIn(url: URL?, completion: @escaping (Bool) -> Void) {
