+++ iOS/issue-tracker/issue-tracker/ViewController/IssueListViewController.swift
@@ -30,7 +30,7 @@ class IssueListViewController: UIViewController {
         configureTableView()
     }
     
-    func configureViewModel() {
+    private func configureViewModel() {
         issueListViewModel = IssueListViewModel(issueListUseCase: DefaultIssueListUseCase())
     }
