+++ iOS/issue-tracker/issue-tracker/ViewController/MilestoneListViewController.swift
@@ -43,6 +43,7 @@ class MilestoneListViewController: UIViewController {
     @objc func showNewMilestoneView() {
         // 마일스톤 추가 화면으로 전환하기
     }
+    
 }
 
 extension MilestoneListViewController: UITableViewDataSource, UITableViewDelegate {
